==> Script/R_Script/M2/AV_AP_ARG/00-main.R <==
library(tidyverse)
library(tidytree)
library(ape)
library(ggtree) 

source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/01-02_alt.R') 
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/03-Taxonomy_parser.R')
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/04-Taxo_Join.R') 
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/05-Matrix_Preparator.R')
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/06_Matrix_Generator.R') 
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/07-Matrix_Plot.R') 
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/08-Taxo_Filter_Plot.R') 
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/09-Tree_Pruner.R') 
source('W:/ninon-species/script/Script_M2/AV_AP_ARG/Matrix/10-Tree_Plot.R')
